requirements.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.6
aiofiles>=23.2.1
jinja2>=3.1.3

python-dotenv>=1.0.0
pydantic-settings>=2.1.0
loguru>=0.7.2

pillow>=10.0.0
pillow-heif>=0.18.0
